Subsumed CMS type collections under a common name
Both content pages as well as glossary pages can render the same set of
elements. In order to save us from synchronizing this set between
these page types, this commit moves common building blocks to a common
location.

This also fixes a bug where attempting to render the "Lasub" target's
production environment crashed the renderer.

The issue was that boxes appearing on a content page were handled, but
boxes that would appear on a glossary page were not.


Former-commit-id: a9e3474dd37170a60ce09a9cea05403456aedd71