Simplified database code
This commit simplifies database-related code by exposing the database
itself as opposed to wrapping it in functions that mostly do the same
thing.

Further, this commit writes output data into a single directory that
is fully disjoint from the source tree, named `dist`.

`dist` is considered to be the default name for such a directory in
NodeJS projects.


Former-commit-id: ba67205fbadeb6f71842756f0186321ce13c3a0d