Made distribution directory configurable
Internal discussions have made it apparent that different developers
adhere to different standards.

In order to make everyone's development experience marvellous, this
commit adds a `compliance` option that can be set to "nodejs" if a
development experience typical to NodeJS projects is desired.

As of now, this controls the name of what's typically known as the
"dist" directory, which systemx will continue to default to "web".

This mechanism can be further extended to allow for true out-of-source
builds.


Former-commit-id: 862efe56f922260845c45afd68a62d2d52326d32