Made the integration tests actually test something
Before the integration tests were mostly useless, since the pipeline
would just skip over the erroneous entry. Now there is a separate script
`test:integration` which tries to build a release but instead of
printing an error message it just rethrows the error leading to an early
exit which the CI Pipeline should detect and interpret as an error.